Cast line points into int due to overload resolution issue in OpenCV

+++ misc/objdet_tools.py
@@ -228,12 +228,12 @@ def project_detections_into_bev(bev_map, detections, configs, color=[]):
         bev_corners[3, 1] = y + w / 2 * sin_yaw + l / 2 * cos_yaw
         
         # draw object as box
-        corners_int = bev_corners.reshape(-1, 1, 2).astype(int)
+        corners_int = bev_corners.reshape(-1, 1, 2).astype(np.int32)
         cv2.polylines(bev_map, [corners_int], True, color, 2)
 
         # draw colored line to identify object front
         corners_int = bev_corners.reshape(-1, 2)
-        cv2.line(bev_map, (corners_int[0, 0], corners_int[0, 1]), (corners_int[3, 0], corners_int[3, 1]), (255, 255, 0), 2)
+        cv2.line(bev_map, (int(corners_int[0, 0]), int(corners_int[0, 1])), (int(corners_int[3, 0]), int(corners_int[3, 1])), (255, 255, 0), 2)
